+++ action/action-dsl/src/main/java/com/powsybl/action/dsl/ast/ActionExpressionEvaluator.java
@@ -209,7 +209,7 @@ public class ActionExpressionEvaluator extends ExpressionEvaluator implements Ac
 
     @Override
     public Object visitIsOverloaded(IsOverloadedNode isOverloadedNode, Void arg) {
-        float limitReduction = isOverloadedNode.getLimitReduction();
+        double limitReduction = isOverloadedNode.getLimitReduction();
 
         // Iterate over all the branch Ids to be sure that all the branches exist in the network
         return isOverloadedNode.getBranchIds().stream()
@@ -219,7 +219,7 @@ public class ActionExpressionEvaluator extends ExpressionEvaluator implements Ac
 
     @Override
     public Object visitAllOverloaded(AllOverloadedNode allOverloadedNode, Void arg) {
-        float limitReduction = allOverloadedNode.getLimitReduction();
+        double limitReduction = allOverloadedNode.getLimitReduction();
 
         // Iterate over all the branch Ids to be sure that all the branches exist in the network
         return allOverloadedNode.getBranchIds().stream()
